changes on redesign UI
